requirements.txt
gradio
# datasets is already present
datasets
numpy
python-dotenv
openai
soundfile 